Separate Engine and Game
Engine now takes care of all mod loading. Engine will fire EngineEvent as it initializes, Mods should listen to these events to register. These are not tested since the GLAnimatedTexture error is not fixed.

Also, EngineDummyContainer is moved to EngineClient since for now it needs to handle registration of renderers and textures. Once we start making client and server proxies we can handle this correctly